Move javascript out of /assets folder
