+++ src/data/ajlamps.js
