+++ src/modules/lamps.js
